refactored about generator and profile cards
