--- data/people/adam-freedman.md
@@ -1,1 +1,0 @@
-[Adam's personal website :octicons-link-external-24:](https://ahfgenomics.com/){:target="_blank"}
