+++ docs/assets/stylesheets/cards.css
@@ -1,0 +1,189 @@
+/*------------------------------------------------------*/
+/* Profile cards */
+.profile-card, .profile-card * {
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+.profile-cards {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  /* column-gap: 5vw; */
+  row-gap: 1.5em;
+  /* margin-bottom: 2em; */
+  justify-content: center;
+}
+.profile-row {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: nowrap;
+  column-gap: 5vw;
+  /* row-gap: 1.5em; */
+  justify-content: center;
+}
+.profile-card {
+  background: #f3e5d8;
+  border: 1px solid #000000;
+  border-radius: 10px;
+  padding: 0.25em 0.75em;
+  margin: 0.5em 0;
+  display: flex;
+  flex-direction: column;
+  flex: 0 1 26vw;
+  max-width: 26vw;
+  min-width: 400px;
+  box-sizing: border-box;
+}
+.profile-card-header {
+  font-size: 1.25em; 
+  color: #db6d6d;
+  font-weight: bold;
+  margin-bottom: 0.2em;
+  text-align: center;
+}
+.profile-card-main {
+  display: flex;
+  flex-direction: row;
+  align-items: flex-start;
+  gap: 1.7em; /* slightly wider gap between image and text */
+}
+.profile-imgcol {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  /* min-width: 170px; */
+  max-width: 30vw;
+  flex: 0 0 180px;
+}
+.profile-pic {
+  border-radius: 10px;
+  width: 8vw;
+  max-width: 195px;
+  min-width: 120px;
+  height: auto;
+  object-fit: cover;
+  margin-bottom: 0.7em;
+  display: block;
+}
+.profile-details {
+  flex: 1;
+  font-size: 0.9em;
+}
+.profile-title {
+  font-size: 0.9em;
+  font-weight: bold;
+  margin-bottom: 0.1em;
+  color: #000000
+}
+.profile-office {
+  font-size: 0.9em;
+  color: #555;
+  margin-bottom: 0.6em;
+}
+.profile-bio {
+  margin-bottom: 0.9em;
+  font-size: 0.9em;
+  color: #000000
+}
+.profile-links {
+  display: flex;
+  justify-content: center;
+  gap: 0.8em;
+  font-size: 1.6em;
+  width: fit-content;
+  margin: 0.1em auto 0 auto;
+}
+.profile-links a {
+  color: #222;
+  margin-right: 0;
+  text-decoration: none;
+}
+.profile-links a:hover {
+  color: #c03a29;
+}
+
+.profile-links .contact-icon svg,
+.profile-links .contact-icon img {
+  width: 32px;
+  height: 32px;
+  vertical-align: middle;
+}
+
+/* Responsive for profile cards only */
+@media (max-width: 1100px) {
+  .profile-row {
+    flex-direction: column;
+    align-items: center;
+    gap: 1.5em;
+  }
+  .profile-card {
+    width: 95%;
+    max-width: 80%;
+    min-width: 220px;
+    flex: none;
+  }
+  .profile-pic { width: 36vw; min-width: 70px; max-width: 180px; }
+}
+
+/*------------------------------------------------------*/
+/* Alumni cards */
+
+.alumni-cards {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 2em 2.5vw;
+  margin: 2em;
+}
+
+.alumni-card {
+  background: #f5eade;
+  border: 1px solid #333333;
+  border-radius: 10px;
+  padding: 1em 1.2em 1em 1.2em;
+  margin: 0.5em 0;
+  min-width: 290px;
+  max-width: 340px;
+  flex: 1 1 280px;
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  font-size: 0.85em;
+}
+
+.alumni-name, .alumni-link {
+  font-weight: bold;
+}
+
+.alumni-link {
+  color: #db6d6d !important;
+  text-decoration: none;
+}
+
+.alumni-years {
+  color: #444;
+  margin-bottom: 0.5em;
+}
+
+.alumni-profile {
+  font-size: 0.85em;
+}
+
+/* Responsive for alumni cards only */
+@media (max-width: 1300px) {
+  .alumni-cards {
+    grid-template-columns: repeat(2, 1fr);
+    justify-items: center; 
+  }
+  .alumni-card { width: 40%; }
+}
+@media (max-width: 1000px) {
+  .alumni-card { width: 80% }
+  .alumni-cards  { flex-direction: column; }
+}
+@media (max-width: 750px) {
+  .alumni-cards {
+    grid-template-columns: 1fr;
+  }
+}
+
