+++ docs/people/adam-freedman.md
@@ -11,17 +11,24 @@ hide:
   </span>
 </h1>
 
-[Adam's personal website :octicons-link-external-24:](https://ahfgenomics.com/){:target="_blank"}
+
+
+
+---
+
+## Where to find:
+
+:material-web: [Website :octicons-link-external-24:](https://ahfgenomics.com/) &nbsp; :simple-github: [GitHub :octicons-link-external-24:](https://github.com/adamfreedman) &nbsp; :scholar-100: [Scholar :octicons-link-external-24:](https://scholar.google.com/citations?hl=en&user=6Bt9l1wAAAAJ&view_op=list_works&sortby=pubdate)
 
 ---
 
 ## Pages authored by Adam Freedman
 
- - [[Workshop] Harvard Informatics R ggplot Workshop](../workshops/short-trainings/r-ggplot/R_ggplot_revised.md)
- - [[Workshop] Harvard Informatics R Tidyverse Workshop](../workshops/short-trainings/r-tidyverse/R_tidyverse_revised.md)
+ - [[Tutorial] Genome annotation](../resources/Tutorials/how-to-annotate-a-genome.md)
  - [[Workshop] Biotips Day 1: Bioinformatics command line tools and file formats](../workshops/biotips/Biotips-workshop-Day1.md)
  - [[Workshop] Harvard Informatics Healthy Habits for Data Science Workshop](../workshops/healthy-habits/healthy_habits_day4.md)
- - [[Workshop] Introduction to R Part 3: ggplot](../workshops/intro-r/R-workshop-Part3.md)
  - [[Workshop] Introduction to R Part 1: R Basics](../workshops/intro-r/R-workshop-Part1.md)
  - [[Workshop] Introduction to R Part 2: Data manipulation with tidyverse](../workshops/intro-r/R-workshop-Part2.md)
- - [[Tutorial] Genome annotation](../resources/Tutorials/how-to-annotate-a-genome.md)
+ - [[Workshop] Introduction to R Part 3: ggplot](../workshops/intro-r/R-workshop-Part3.md)
+ - [[Workshop] Harvard Informatics R ggplot Workshop](../workshops/short-trainings/r-ggplot/R_ggplot_revised.md)
+ - [[Workshop] Harvard Informatics R Tidyverse Workshop](../workshops/short-trainings/r-tidyverse/R_tidyverse_revised.md)
