+++ docs/people/gregg-thomas.md
@@ -21,27 +21,34 @@ To foster best practices and reproducibility in computational biology he also te
 
 Outside of the lab, he has a passion for dogs and animals in general. He also enjoys being active and doing puzzles with his partner!
 
-See [Gregg's personal website :octicons-link-external-24:](https://gwct.bio/){:target="_blank"} for his CV, publication list, and software.
+See [Gregg's website :octicons-link-external-24:](https://gwct.bio/){:target="_blank"} for his CV, publication list, and software.
+
+
+---
+
+## Where to find:
+
+:material-web: [Website :octicons-link-external-24:](https://gwct.bio/) &nbsp; :simple-github: [GitHub :octicons-link-external-24:](https://github.com/gwct) &nbsp; :scholar-100: [Scholar :octicons-link-external-24:](https://scholar.google.com/citations?hl=en&user=4pydfEUAAAAJ&view_op=list_works&sortby=pubdate)
 
 ---
 
 ## Pages authored by Gregg Thomas
 
- - [[Workshop] Harvard Informatics R ggplot Workshop](../workshops/short-trainings/r-ggplot/R_ggplot_revised.md)
- - [[Workshop] Harvard Informatics R Tidyverse Workshop](../workshops/short-trainings/r-tidyverse/R_tidyverse_revised.md)
+ - [[Tutorial] Adding an outgroup to a whole genome alignment](../resources/Tutorials/add-outgroup-to-whole-genome-alignment-cactus.md)
+ - [[Tutorial] Adding a genome to a whole genome alignment](../resources/Tutorials/add-to-whole-genome-alignment-cactus.md)
+ - [[Tutorial] Installing command line software with conda/mamba](../resources/Tutorials/installing-command-line-software-conda-mamba.md)
+ - [[Tutorial] Pangenome inference with Cactus-minigraph](../resources/Tutorials/pangenome-cactus-minigraph.md)
+ - [[Tutorial] Replacing a genome in a whole genome alignment](../resources/Tutorials/replace-genome-whole-genome-alignment-cactus.md)
+ - [[Tutorial] Whole genome alignment with Cactus](../resources/Tutorials/whole-genome-alignment-cactus.md)
  - [[Workshop] Biotips Day 1: Bioinformatics command line tools and file formats](../workshops/biotips/Biotips-workshop-Day1.md)
  - [[Workshop] Biotips Day 2: Bioinformatics command line tools and file formats](../workshops/biotips/Biotips-workshop-Day2.md)
- - [[Workshop] Biotips Day 4: Shell scripting](../workshops/biotips/Biotips-workshop-Day4.md)
  - [[Workshop] Biotips Day 3: VCF files and shell scripts](../workshops/biotips/Biotips-workshop-Day3.md)
+ - [[Workshop] Biotips Day 4: Shell scripting](../workshops/biotips/Biotips-workshop-Day4.md)
  - [[Workshop] Harvard Informatics Healthy Habits for Data Science Workshop](../workshops/healthy-habits/healthy_habits_day2.md)
- - [[Workshop] Introduction to R Part 3: ggplot](../workshops/intro-r/R-workshop-Part3.md)
  - [[Workshop] Introduction to R Part 1: R Basics](../workshops/intro-r/R-workshop-Part1.md)
  - [[Workshop] Introduction to R Part 2: Data manipulation with tidyverse](../workshops/intro-r/R-workshop-Part2.md)
- - [[Tutorial] Adding an outgroup to a whole genome alignment](../resources/Tutorials/add-outgroup-to-whole-genome-alignment-cactus.md)
- - [[Tutorial] Whole genome alignment with Cactus](../resources/Tutorials/whole-genome-alignment-cactus.md)
- - [[Tutorial] Installing command line software with conda/mamba](../resources/Tutorials/installing-command-line-software-conda-mamba.md)
- - [[Tutorial] Adding a genome to a whole genome alignment](../resources/Tutorials/add-to-whole-genome-alignment-cactus.md)
- - [[Tutorial] Replacing a genome in a whole genome alignment](../resources/Tutorials/replace-genome-whole-genome-alignment-cactus.md)
- - [[Tutorial] Pangenome inference with Cactus-minigraph](../resources/Tutorials/pangenome-cactus-minigraph.md)
- - [[Maintainer] Harvard Informatics Software](../software/index.md)
+ - [[Workshop] Introduction to R Part 3: ggplot](../workshops/intro-r/R-workshop-Part3.md)
+ - [[Workshop] Harvard Informatics R ggplot Workshop](../workshops/short-trainings/r-ggplot/R_ggplot_revised.md)
+ - [[Workshop] Harvard Informatics R Tidyverse Workshop](../workshops/short-trainings/r-tidyverse/R_tidyverse_revised.md)
  - [[Maintainer] Computing glossary](../resources/glossary.md)
+ - [[Maintainer] Harvard Informatics Software](../software/index.md)
