+++ docs/people/katya-osipova.md
@@ -1,0 +1,24 @@
+---
+hide:
+    - toc
+---
+
+<h1 class="profile-header">
+  <img src="/img/people/katyaosipova.jpg" alt="Katya Osipova">
+  <span class="profile-name">
+    Katya Osipova<br>
+    <span class="profile-title">Postdoctoral Fellow</span>
+  </span>
+</h1>
+
+
+
+
+---
+
+## Where to find:
+
+:material-web: [Website :octicons-link-external-24:](https://osipovarev.github.io/) &nbsp; :simple-github: [GitHub :octicons-link-external-24:](https://github.com/osipovarev) &nbsp; :scholar-100: [Scholar :octicons-link-external-24:](https://scholar.google.com/citations?hl=en&user=OhvZN8cAAAAJ&view_op=list_works&sortby=pubdate)
+
+---
+
