+++ docs/people/lei-ma.md
@@ -13,17 +13,24 @@ hide:
 
 
 
+
+---
+
+## Where to find:
+
+:simple-github: [GitHub :octicons-link-external-24:](https://github.com/microlei)
+
 ---
 
 ## Pages authored by Lei Ma
 
- - [[Workshop] Harvard Informatics R ggplot Workshop](../workshops/short-trainings/r-ggplot/R_ggplot_revised.md)
- - [[Workshop] Harvard Informatics R Tidyverse Workshop](../workshops/short-trainings/r-tidyverse/R_tidyverse_revised.md)
  - [[Workshop] Biotips Day 1: Bioinformatics command line tools and file formats](../workshops/biotips/Biotips-workshop-Day1.md)
- - [[Workshop] Biotips Day 4: Shell scripting](../workshops/biotips/Biotips-workshop-Day4.md)
  - [[Workshop] Biotips Day 3: VCF files and shell scripts](../workshops/biotips/Biotips-workshop-Day3.md)
- - [[Workshop] Harvard Informatics Healthy Habits for Data Science Workshop](../workshops/healthy-habits/healthy_habits_day3.md)
+ - [[Workshop] Biotips Day 4: Shell scripting](../workshops/biotips/Biotips-workshop-Day4.md)
  - [[Workshop] Harvard Informatics Healthy Habits for Data Science Workshop](../workshops/healthy-habits/healthy_habits_day1.md)
- - [[Workshop] Introduction to R Part 3: ggplot](../workshops/intro-r/R-workshop-Part3.md)
+ - [[Workshop] Harvard Informatics Healthy Habits for Data Science Workshop](../workshops/healthy-habits/healthy_habits_day3.md)
  - [[Workshop] Introduction to R Part 1: R Basics](../workshops/intro-r/R-workshop-Part1.md)
  - [[Workshop] Introduction to R Part 2: Data manipulation with tidyverse](../workshops/intro-r/R-workshop-Part2.md)
+ - [[Workshop] Introduction to R Part 3: ggplot](../workshops/intro-r/R-workshop-Part3.md)
+ - [[Workshop] Harvard Informatics R ggplot Workshop](../workshops/short-trainings/r-ggplot/R_ggplot_revised.md)
+ - [[Workshop] Harvard Informatics R Tidyverse Workshop](../workshops/short-trainings/r-tidyverse/R_tidyverse_revised.md)
