+++ docs/people/nathan-weeks.md
@@ -13,10 +13,12 @@ hide:
 
 
 
+
+
 ---
 
 ## Pages authored by Nathan Weeks
 
- - [[Workshop] Biotips Day 4: Shell scripting](../workshops/biotips/Biotips-workshop-Day4.md)
- - [[Workshop] Biotips Day 3: VCF files and shell scripts](../workshops/biotips/Biotips-workshop-Day3.md)
  - [[Tutorial] How to download sequencing data from the Bauer Core](../resources/Tutorials/how-can-i-download-my-sequencing-data.md)
+ - [[Workshop] Biotips Day 3: VCF files and shell scripts](../workshops/biotips/Biotips-workshop-Day3.md)
+ - [[Workshop] Biotips Day 4: Shell scripting](../workshops/biotips/Biotips-workshop-Day4.md)
