+++ etc/old-card-styles.css
@@ -1,0 +1,178 @@
+/*------------------------------------------------------*/
+/* Profile cards */
+
+.card-row {
+  min-height:275px;
+  height: 30vh;
+  max-height: 300px;
+}
+
+.alum-card-row {
+  height: 15vh;
+  font-size: 0.85em;
+  max-height: 175px;
+}
+
+.card-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  vertical-align: middle;
+  border: 1px solid #333333;
+  background-color: #f3e5d8;
+  -webkit-border-radius:5px;
+  -moz-border-radius:5px;
+  border-radius:5px;
+  height:100%;
+  /* padding: 5px; */
+  font-size: 0.85em;
+}
+
+.alum-card-container {
+  padding: 10px;
+  overflow-y: scroll;
+  -ms-overflow-style: none;  /* IE and Edge */
+  scrollbar-width: none;  /* Firefox */
+}
+
+/* Hide scrollbar for Chrome, Safari and Opera */
+.alum-card-container::-webkit-scrollbar {
+  display: none;
+}
+
+.alum-card-container-no-profile {
+  padding: 10px;
+  height: 30%;
+  overflow-y: scroll;
+  -ms-overflow-style: none;  /* IE and Edge */
+  scrollbar-width: none;  /* Firefox */
+}
+
+/* Hide scrollbar for Chrome, Safari and Opera */
+.alum-card-container-no-profile::-webkit-scrollbar {
+  display: none;
+}
+
+.alum-name {
+  font-size: 1.1em;
+  font-weight: bold;
+}
+
+.card-name-container {
+  height: 14%;
+}
+
+.card-name {
+  display: flex !important;
+  justify-content: center !important;
+  align-items: center !important;
+  font-size: 1.5em;
+}
+
+.card-name-sep-div {
+  height: 1.5%;
+}
+
+.card-content-container {
+  /* height: 25vh; */
+  height: 77%;
+  }
+
+.card-img-container {
+  height: 100%;
+}
+
+.card-img {
+  width: auto !important;
+  /* height: 150px !important; */
+  /* max-width: 100% !important; */
+  /* height: 20vh; */
+  height: auto !important;
+  max-height: 100% !important;
+  -webkit-border-radius:5px;
+  -moz-border-radius:5px;
+  border-radius:5px;
+}
+
+.card-content {
+  height: 100%;
+  overflow-wrap: break-word;
+  overflow-y: scroll;
+  -ms-overflow-style: none;  /* IE and Edge */
+  scrollbar-width: none;  /* Firefox */
+}
+
+/* Hide scrollbar for Chrome, Safari and Opera */
+.card-content::-webkit-scrollbar {
+  display: none;
+}
+
+.card-sep-div {
+  height: 3vh;
+  /* height: 7.5% */
+}
+
+.icon-note {
+  width: 100%;
+  color: #999999;
+  font-size: 0.6em;
+  align-content: left;
+}
+
+@media (max-width: 959px) {
+  .card-row, .alum-card-row {
+    height: unset !important;
+    max-height: unset;
+    flex-direction: column;
+    justify-content: center;
+  }
+  .card-container {  
+    display: block !important;
+    width: 70vw !important;
+  }
+
+  .card-name-margin {
+    width: 0 !important;
+    display: none !important;
+  }
+  .card-name {
+    width: 100% !important;   
+  }
+
+  .card-content-container {
+    height: unset !important;
+    flex-direction: column;
+    justify-content: center;
+  }
+
+  .card-img-container {
+    display: flex !important;
+    justify-content: center !important;
+    align-items: center !important;
+    flex-direction: column !important;
+    width: 100% !important;
+  }
+
+  .card-content {
+    display: block !important;
+    width: 100% !important;
+    padding: 10px;
+  }
+
+  .card-img {
+    width: auto !important;
+    height: 150px !important;
+  } 
+
+  .card-margin-inner {
+    display: block !important;
+    width: 100vw !important;
+    height: 3vh;
+    display: none;
+  }
+  .card-margin-outer {
+    display: block !important;
+    width: 10vw !important;
+  }
+}
+/* Responsive styles */
